feat(v0batchuserscontroller.js): combined read all controller with read w/out query params
both as a lean practice as well as fix bug where read all consumes the query and errors for query
params
